+++ app/(routes)/login/LoginForm.tsx
@@ -3,7 +3,7 @@
 import DFormFieldComponent from "@/components/DFormFieldComponent";
 import {Button} from "@/components/ui/button";
 import {Form} from "@/components/ui/form";
-import {login} from "@/lib/authApi";
+import {login} from "@/lib/api/authApi";
 import {manualSignIn} from "@/lib/firebase/utils";
 import {zodResolver} from "@hookform/resolvers/zod";
 import {useMutation} from "@tanstack/react-query";
