+++ app/(routes)/signup/SignUpForm.tsx
@@ -3,7 +3,7 @@
 import DFormFieldComponent from "@/components/DFormFieldComponent";
 import {Button} from "@/components/ui/button";
 import {Form} from "@/components/ui/form";
-import {signUp} from "@/lib/authApi";
+import {signUp} from "@/lib/api/authApi";
 import {createUserWithEmail} from "@/lib/firebase/utils";
 import {zodResolver} from "@hookform/resolvers/zod";
 import {useMutation} from "@tanstack/react-query";
